notebook/qa_over_docs/requirements.txt
langchain
pdfminer.six
faiss-cpu
tiktoken
openai